Move ClassDefinitionManager into Pimcore\Model\DataObject\ClassDefinition
